+++ include/Ace/Entity.h
@@ -6,7 +6,7 @@
 namespace ace
 {
 
-    class Entity
+    class Entity final
     {
         EntityManager::EntityHandle* m_handle;
 
@@ -168,8 +168,8 @@ namespace ace
         }
 
         /**
-            @brief Queries the entity static status.
-            @return True if the entity is static. (Static entities transforms are not recalculated on each frame)
+            @brief Queries the entity static status. Static entities transforms are recalculated only if their parents models change.
+            @return True if the entity is static.
         */
         inline bool IsStatic() const
         {
@@ -212,15 +212,6 @@ namespace ace
         void SetParent(Entity& parent)
         {
             parent.AddChild(*this);
-        }
-
-        /**
-            @brief Set static status of the entity.
-            @param[in] isStatic True if the entity should be made static.
-        */
-        void SetStatic(const bool isStatic)
-        {
-            if (m_handle) m_handle->SetStatic(isStatic);
         }
 
         /**