+++ src/Ace/Scene.cpp
@@ -1,44 +1,36 @@
 #include <Ace/Scene.h>
 
 #include <Ace/Entity.h>
-#include <Ace/Matrix4.h>
 #include <Ace/GraphicsDevice.h>
+#include <Ace/Log.h>
+#include <Ace/Matrix4.h>
 #include <Ace/SpriteManager.h>
-
-#include <Ace/Platform.h>
-
-#if ACE_DEBUG
-    #include <Ace/Log.h>
-#endif
 
 namespace ace
 {
-
-    void ComputeMatrices(Entity& entity, const math::Matrix4& parentModel)
+    void ComputeMatrices(EntityHandle* handle, const math::Matrix4& parentModel, bool hasParentModelChanged)
     {
-        if (!entity)
+        if (!handle)
         {
-            #if ACE_DEBUG
-                Logger::LogDebug("ComputeMatrices null entity");
-            #endif
-
+            Logger::Log(Logger::Priority::Warning, "Scene: Update: Skipped null entity");
             return;
         }
 
-        if (!entity->transform.GetModelStatus())
+        // Changing parent model overrides entity isStatic
+        if (hasParentModelChanged || (handle->transform.HasModelChanged() && !handle->IsStatic()))
         {
-            entity->transform.SetModel(
-                math::MakeScaling(entity->transform.GetScale()) *
-                math::ToMatrix<4u>(entity->transform.GetRotation()) *
-                math::MakeTranslation(entity->transform.GetPosition()) * parentModel
+            handle->transform.SetModel(
+                math::MakeScaling(handle->transform.GetScale()) *
+                math::ToMatrix<4u>(handle->transform.GetRotation()) *
+                math::MakeTranslation(handle->transform.GetPosition()) * parentModel
             );
+            hasParentModelChanged = true;
         }
 
-        const UInt32 count = entity.ChildCount();
+        const UInt32 count = handle->ChildCount();
         for (UInt32 i = 0u; i < count; ++i)
         {
-            Entity child = entity.GetChild(i);
-            ComputeMatrices(child, entity->transform.GetModel());
+            ComputeMatrices(handle->GetChild(i), handle->transform.GetModel(), hasParentModelChanged);
         }
     }
 
@@ -78,7 +70,7 @@ namespace ace
 
     void Scene::Update()
     {
-        ComputeMatrices(*m_root, math::s_identity4);
+        ComputeMatrices(*m_root, math::s_identity4, false);
     }
 
 }