+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+
+
+## 1.11.2017
+
+### Changed
+
+- Improved support for static entities
+
+
+
 ## 30.10.2017
 
 ### Added
@@ -9,6 +19,33 @@
 - Animation Example
 - Added TotalTime method for Time
 - Added GetScale method for Sprite
+
+
+
+## 29.10.2017
+
+### Added
+
+- Proper CollisionExample
+
+### Changed
+
+- Collidables work when used without ECS
+
+
+
+## 27.10.2017
+
+### Added
+
+- Support for static entities
+
+### Changed
+
+- Transform members are private
+- Transform stores status if it has been modified
+
+
 
 ## 13.10.2017
 
@@ -25,14 +62,16 @@
 - BaseComponent, fixed constants
 - Path, added operators and absolute path
 
+
+
 ## 29.09.2017
+
 ### Added
 
 - Addded CHANGELOG.md
 - Added Vector4 operators
 - Added GetParent to EntityHandle
 - Added LookAt and Identity to Quaternation
-- 
 
 ### Changed